+++ react-webapp/src/WeekMenu/RecipeList.js
@@ -6,6 +6,7 @@ import MessageAlert from '../MessageAlert';
 import {errorMessage} from '../common/Util';
 import {get} from "../services/ApiService";
 import HomeContent from '../home/HomeContent';
+import { confirmDialog } from '../common/ConfirmDialog';
 
 class RecipeList extends Component {
   constructor(props) {
@@ -16,21 +17,18 @@ class RecipeList extends Component {
 
   async componentDidMount() {
     let jwt = this.state.jwt;
-    console.log("JWT: ", jwt);
     if (jwt) {
       try {
-        const data = await get('week-menu/v2/recipe', true, false, jwt)
+        let data = await get('week-menu/v2/recipe', true, false, jwt)
         if (data) {
-          let jsonData = JSON.parse(data);
-          this.setState({isLoading: false, recipes: jsonData});
-        } else {
-          this.setState({isLoading: false, displayError: errorMessage(data)});
+          data = JSON.parse(data);
+          if (Array.isArray(data)) {
+            this.setState({isLoading: false, recipes: data});
+          } else {
+            this.setState({isLoading: false, displayError: errorMessage(data)});
+          }
         }
       } catch (error) {
-        if (error.status === 401 || error.status === 403) {
-          this.props.onRemoveAuthentication();
-          this.props.history.push('/');
-        }
         this.setState({isLoading: false, displayError: errorMessage(error)});
       }
     }
@@ -44,18 +42,26 @@ class RecipeList extends Component {
     }
   }
 
-  async remove(id) {
-    await fetch(`/api/week-menu/v2/recipe/${id}`, {
-      method: 'DELETE',
-      headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json'
-      },
-      credentials: 'include'
-    }).then(() => {
-      let recipes = [...this.state.recipes].filter(i => i._id !== id);
-      this.setState({recipes: recipes});
-    });
+  async remove(recipe) {
+    let confirm = await confirmDialog(`Delete Recipe ${recipe.name}`, "Are you sure you want to delete this?", "Delete Recipe");
+    if (confirm) {
+      let id = recipe.id;
+      let jwt = this.state.jwt;
+      await fetch(`/api/week-menu/v2/recipe/${id}`, {
+        method: 'DELETE',
+        headers: {
+          'Authorization': jwt
+        },
+        credentials: 'include'
+      }).then((err) => {
+        if (err.status !== 200) {
+          this.setState({displayError: errorMessage(err)});
+        } else {
+          let recipes = [...this.state.recipes].filter(i => i._id !== id);
+          this.setState({recipes: recipes});
+        }
+      });
+    }
   }
 
   render() {
@@ -73,7 +79,7 @@ class RecipeList extends Component {
         <td>
           <ButtonGroup>
             <Button size="sm" color="primary" tag={Link} to={"/recipes/" + recipe._id}>Edit</Button>
-            <Button size="sm" color="danger" onClick={() => this.remove(recipe._id)}>Delete</Button>
+            <Button size="sm" color="danger" onClick={() => this.remove({'id': recipe._id, 'name': recipe.name})}>Delete</Button>
           </ButtonGroup>
         </td>
       </tr>
